requirements.txt
## The following requirements were added by pip freeze:
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
fastapi==0.61.1
google-api-core==1.22.2
google-auth==1.22.0
google-cloud-core==1.4.1
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
h11==0.10.0
idna==2.10
multidict==4.7.6
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.6.1
python-gitlab==2.5.0
pytz==2020.1
requests==2.24.0
rsa==4.6
sentry-sdk==0.18.0
six==1.15.0
starlette==0.13.6
urllib3==1.25.10
uvicorn==0.12.0
yarl==1.6.0
